context-retrieval-state/projects/project-2.4-meeting-notes-assistant/requirements.txt
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
pydantic>=2.7.0
python-dotenv>=1.0.1
